Re-written code for UV selection sets
re-written the creation of the selection sets. It's now a bit faster
and code is more structured.

New functionality: Check and Fix UVs
UV position is now verified if it is inside a UDIM if not the points
are slightly moved inside.
Polygons which spread over two or more UDIMs are now selected and a
warning message is shown.